+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## unreleased
 
+- feat: added edit subcommand as VS Code API, removed prettify
 - feat: added prettify --line N for in-file table detection and reformat
 
 ## v0.2.1 (2026-06-13)